cria modulo libs
