+++ idessem/dessem/__init__.py
@@ -1,37 +1,34 @@
-
 # Inclui os membros
 
 from .avl_altqueda import AvlAltQueda  # noqa
 from .avl_desvfpha import AvlDesvFpha  # noqa
-from .avl_estatfpha import AvlEstatFpha    # noqa
-from .avl_fpha1 import AvlFpha1    # noqa
-from .avl_fpha2 import AvlFpha2   # noqa
-from .avl_fpha3 import AvlFpha3   # noqa
-from .des_log_relato import DesLogRelato   # noqa
-from .dessemarq import DessemArq    # noqa
-from .dessopc import Dessopc    # noqa
-from .entdados import Entdados    # noqa
-from .hidr import Hidr    # noqa
-from .log_matriz import LogMatriz    # noqa
-from .operuh import Operuh    # noqa
-from .operut import Operut    # noqa
-from .pdo_aval_qmaxusih import PdoAvalQmaxUsih    # noqa
-from .pdo_eco_fcfcortes import PdoEcoFcfCortes    # noqa
-from .pdo_eco_usih_conj import PdoEcoUsihConj    # noqa
-from .pdo_eco_usih_polin import PdoEcoUsihPolin    # noqa
-from .pdo_eco_usih import PdoEcoUsih    # noqa
-from .pdo_eolica import PdoEolica    # noqa
-from .pdo_hidr import PdoHidr    # noqa
-from .pdo_inter import PdoInter    # noqa
-from .pdo_oper_term import PdoOperTerm    # noqa
-from .pdo_oper_tviag_calha import PdoOperTviagCalha    # noqa
-from .pdo_oper_uct import PdoOperUct    # noqa
-from .pdo_operacao import PdoOperacao    # noqa
-from .pdo_reserva import PdoReserva    # noqa
-from .pdo_sist import PdoSist    # noqa
-from .pdo_term import PdoTerm    # noqa
-from .polinjus import Polinjus    # noqa
-from .renovaveis import Renovaveis    # noqa
-from .termdat import Term    # noqa
-from .uch import Uch    # noqa
-from .vazaolateral import VazaoLateral    # noqa
+from .avl_estatfpha import AvlEstatFpha  # noqa
+from .avl_fpha1 import AvlFpha1  # noqa
+from .avl_fpha2 import AvlFpha2  # noqa
+from .avl_fpha3 import AvlFpha3  # noqa
+from .des_log_relato import DesLogRelato  # noqa
+from .dessemarq import DessemArq  # noqa
+from .dessopc import Dessopc  # noqa
+from .entdados import Entdados  # noqa
+from .hidr import Hidr  # noqa
+from .log_matriz import LogMatriz  # noqa
+from .operuh import Operuh  # noqa
+from .operut import Operut  # noqa
+from .pdo_aval_qmaxusih import PdoAvalQmaxUsih  # noqa
+from .pdo_eco_fcfcortes import PdoEcoFcfCortes  # noqa
+from .pdo_eco_usih_conj import PdoEcoUsihConj  # noqa
+from .pdo_eco_usih_polin import PdoEcoUsihPolin  # noqa
+from .pdo_eco_usih import PdoEcoUsih  # noqa
+from .pdo_eolica import PdoEolica  # noqa
+from .pdo_hidr import PdoHidr  # noqa
+from .pdo_inter import PdoInter  # noqa
+from .pdo_oper_term import PdoOperTerm  # noqa
+from .pdo_oper_tviag_calha import PdoOperTviagCalha  # noqa
+from .pdo_oper_uct import PdoOperUct  # noqa
+from .pdo_operacao import PdoOperacao  # noqa
+from .pdo_reserva import PdoReserva  # noqa
+from .pdo_sist import PdoSist  # noqa
+from .pdo_term import PdoTerm  # noqa
+from .renovaveis import Renovaveis  # noqa
+from .termdat import Term  # noqa
+from .uch import Uch  # noqa
