+++ idessem/libs/__init__.py
@@ -1,0 +1,3 @@
+# Inclui os membros
+
+from .usinas_hidreletricas import UsinasHidreletricas  # noqa
