+++ idessem/libs/modelos/usinas_hidreletricas.py
@@ -10,6 +10,8 @@ from typing import Optional
 class HidreletricaCurvaJusante(Register):
     """ """
 
+    __slots__ = []
+
     IDENTIFIER = "HIDRELETRICA-CURVAJUSANTE"
     IDENTIFIER_DIGITS = 26
     LINE = Line(
@@ -68,6 +70,8 @@ class HidreletricaCurvaJusante(Register):
 class HidreletricaCurvaJusantePolinomioPorPartes(Register):
     """ """
 
+    __slots__ = []
+
     IDENTIFIER = "HIDRELETRICA-CURVAJUSANTE-POLINOMIOPORPARTES"
     IDENTIFIER_DIGITS = 45
     LINE = Line(
@@ -124,6 +128,8 @@ class HidreletricaCurvaJusantePolinomioPorPartes(Register):
 
 class HidreletricaCurvaJusantePolinomioPorPartesSegmento(Register):
     """ """
+
+    __slots__ = []
 
     IDENTIFIER = "HIDRELETRICA-CURVAJUSANTE-POLINOMIOPORPARTES-SEGMENTO"
     IDENTIFIER_DIGITS = 54
@@ -289,6 +295,8 @@ class HidreletricaCurvaJusantePolinomioPorPartesSegmento(Register):
 class HidreletricaCurvaJusanteAfogamentoExplicitoUsina(Register):
     """ """
 
+    __slots__ = []
+
     IDENTIFIER = "HIDRELETRICA-CURVAJUSANTE-AFOGAMENTO-EXPLICITO-USINA"
     IDENTIFIER_DIGITS = 53
     LINE = Line(
@@ -331,6 +339,8 @@ class HidreletricaCurvaJusanteAfogamentoExplicitoUsina(Register):
 class HidreletricaCurvaJusanteAfogamentoExplicitoPadrao(Register):
     """ """
 
+    __slots__ = []
+
     IDENTIFIER = "HIDRELETRICA-CURVAJUSANTE-AFOGAMENTO-EXPLICITO-PADRAO"
     IDENTIFIER_DIGITS = 54
     LINE = Line(
@@ -353,3 +363,176 @@ class HidreletricaCurvaJusanteAfogamentoExplicitoPadrao(Register):
     @considera_afogamento.setter
     def considera_afogamento(self, c: str):
         self.data[0] = c
+
+
+class HidreletricaVazaoJusanteInfluenciaDefluencia(Register):
+    """ """
+
+    IDENTIFIER = "HIDRELETRICA-VAZAO-JUSANTE-INFLUENCIA-DEFLUENCIA"
+    IDENTIFIER_DIGITS = 48
+    LINE = Line(
+        [
+            IntegerField(),
+            FloatField(decimal_digits=3),
+            FloatField(decimal_digits=3),
+        ],
+        delimiter=";",
+    )
+
+    @property
+    def codigo_usina_influenciada(self) -> Optional[int]:
+        """
+        O código da usina hidrelétrica influenciada.
+
+        :return: O código da usina
+        :rtype: int | None
+        """
+        return self.data[0]
+
+    @codigo_usina_influenciada.setter
+    def codigo_usina_influenciada(self, c: int):
+        self.data[0] = c
+
+    @property
+    def fator_impacto_turbinamento(self) -> Optional[float]:
+        """
+        O fator de impacto referente ao turbinamento da usina.
+
+        :return: O fator do turbinamento
+        :rtype: float | None
+        """
+        return self.data[1]
+
+    @fator_impacto_turbinamento.setter
+    def fator_impacto_turbinamento(self, c: float):
+        self.data[1] = c
+
+    @property
+    def fator_impacto_vertimento(self) -> Optional[float]:
+        """
+        O fator de impacto referente ao vertimento da usina.
+
+        :return: O fator do vertimento
+        :rtype: float | None
+        """
+        return self.data[2]
+
+    @fator_impacto_vertimento.setter
+    def fator_impacto_vertimento(self, c: float):
+        self.data[2] = c
+
+
+class HidreletricaVazaoJusanteInfluenciaPosto(Register):
+    """ """
+
+    IDENTIFIER = "HIDRELETRICA-VAZAO-JUSANTE-INFLUENCIA-POSTO"
+    IDENTIFIER_DIGITS = 43
+    LINE = Line(
+        [
+            IntegerField(),
+            IntegerField(),
+            FloatField(decimal_digits=3),
+        ],
+        delimiter=";",
+    )
+
+    @property
+    def codigo_usina_influenciada(self) -> Optional[int]:
+        """
+        O código da usina hidrelétrica influenciada.
+
+        :return: O código da usina
+        :rtype: int | None
+        """
+        return self.data[0]
+
+    @codigo_usina_influenciada.setter
+    def codigo_usina_influenciada(self, c: int):
+        self.data[0] = c
+
+    @property
+    def codigo_usina_influenciadora(self) -> Optional[float]:
+        """
+        O código da usina cuja vazão incremental influencia
+        lateralmente.
+
+        :return: O código da usina
+        :rtype: float | None
+        """
+        return self.data[1]
+
+    @codigo_usina_influenciadora.setter
+    def codigo_usina_influenciadora(self, c: float):
+        self.data[1] = c
+
+    @property
+    def fator_impacto(self) -> Optional[float]:
+        """
+        O fator de impacto referente a vazão lateral.
+
+        :return: O fator da vazão lateral
+        :rtype: float | None
+        """
+        return self.data[2]
+
+    @fator_impacto.setter
+    def fator_impacto(self, c: float):
+        self.data[2] = c
+
+
+class HidreletricaVazaoJusanteInfluenciaUsina(Register):
+    """ """
+
+    IDENTIFIER = "HIDRELETRICA-VAZAO-JUSANTE-INFLUENCIA-USINA"
+    IDENTIFIER_DIGITS = 43
+    LINE = Line(
+        [
+            IntegerField(),
+            IntegerField(),
+            FloatField(decimal_digits=3),
+        ],
+        delimiter=";",
+    )
+
+    @property
+    def codigo_usina_influenciada(self) -> Optional[int]:
+        """
+        O código da usina hidrelétrica influenciada.
+
+        :return: O código da usina
+        :rtype: int | None
+        """
+        return self.data[0]
+
+    @codigo_usina_influenciada.setter
+    def codigo_usina_influenciada(self, c: int):
+        self.data[0] = c
+
+    @property
+    def codigo_usina_influenciadora(self) -> Optional[float]:
+        """
+        O código da usina cuja vazão defluente influencia
+        lateralmente.
+
+        :return: O código da usina
+        :rtype: float | None
+        """
+        return self.data[1]
+
+    @codigo_usina_influenciadora.setter
+    def codigo_usina_influenciadora(self, c: float):
+        self.data[1] = c
+
+    @property
+    def fator_impacto(self) -> Optional[float]:
+        """
+        O fator de impacto referente a vazão lateral.
+
+        :return: O fator da vazão lateral
+        :rtype: float | None
+        """
+        return self.data[2]
+
+    @fator_impacto.setter
+    def fator_impacto(self, c: float):
+        self.data[2] = c
